requirements.txt
librosa
noisereduce
scipy
numpy
sounddevice
openai-whisper
torch
rapidfuzz
google-generativeai
python-dotenv
